refactor(frontend): fix naming, structure, and dead code from review
- Rename stepValidator param ServiceMode → serviceMode (camelCase convention)
- Move sumoResponse.ts from components/plots/ to utils/ (pure utility, no JSX)
- De-duplicate snakeToCamelCase: export from functionUtils, import in sumoResponse
- Remove dead @mui/x-data-grid mocks from FunctionList test files (DataGrid removed)
- Rename jobcollection_roundtrip.integration.test.ts → camelCase
- Move tmp_job_collection_import.csv to __fixtures__/ with proper naming

Closes #471
